CMake: Add option to build with WAVM
This adds -DHERA_WAVM option what for now will build WAVM and link hera with it. Because Hera does not uses WAVM it will not do anything useful, but we can check if the build works.

Also the `HERA_WAVM=1` macro is predefined for build with WAVM.

This depends on LLVM 6.0 so you have to install libllvm-6.0-dev. See http://apt.llvm.org.